requirements.txt
pandas
matplotlib
numpy
PyCryptodome